+++ app/controlplane/api/controlplane/v1/shared_message.proto
@@ -41,3 +41,14 @@ message IdentityReference {
     message: "either id or name must be provided, but not both."
   };
 }
+
+
+// ProjectMemberRole defines the roles a member can have in a project
+enum ProjectMemberRole {
+  // Default role for a project member
+  PROJECT_MEMBER_ROLE_UNSPECIFIED = 0;
+  // Admin role for a project member
+  PROJECT_MEMBER_ROLE_ADMIN = 1;
+  // Viewer role for a project member
+  PROJECT_MEMBER_ROLE_VIEWER = 2;
+}
