+++ app/controlplane/api/gen/frontend/controlplane/v1/project.ts
@@ -6,52 +6,14 @@ import { Timestamp } from "../../google/protobuf/timestamp";
 import { Group } from "./group";
 import { OffsetPaginationRequest, OffsetPaginationResponse } from "./pagination";
 import { User } from "./response_messages";
-import { IdentityReference } from "./shared_message";
+import {
+  IdentityReference,
+  ProjectMemberRole,
+  projectMemberRoleFromJSON,
+  projectMemberRoleToJSON,
+} from "./shared_message";
 
 export const protobufPackage = "controlplane.v1";
-
-/** ProjectMemberRole defines the roles a member can have in a project */
-export enum ProjectMemberRole {
-  /** PROJECT_MEMBER_ROLE_UNSPECIFIED - Default role for a project member */
-  PROJECT_MEMBER_ROLE_UNSPECIFIED = 0,
-  /** PROJECT_MEMBER_ROLE_ADMIN - Admin role for a project member */
-  PROJECT_MEMBER_ROLE_ADMIN = 1,
-  /** PROJECT_MEMBER_ROLE_VIEWER - Viewer role for a project member */
-  PROJECT_MEMBER_ROLE_VIEWER = 2,
-  UNRECOGNIZED = -1,
-}
-
-export function projectMemberRoleFromJSON(object: any): ProjectMemberRole {
-  switch (object) {
-    case 0:
-    case "PROJECT_MEMBER_ROLE_UNSPECIFIED":
-      return ProjectMemberRole.PROJECT_MEMBER_ROLE_UNSPECIFIED;
-    case 1:
-    case "PROJECT_MEMBER_ROLE_ADMIN":
-      return ProjectMemberRole.PROJECT_MEMBER_ROLE_ADMIN;
-    case 2:
-    case "PROJECT_MEMBER_ROLE_VIEWER":
-      return ProjectMemberRole.PROJECT_MEMBER_ROLE_VIEWER;
-    case -1:
-    case "UNRECOGNIZED":
-    default:
-      return ProjectMemberRole.UNRECOGNIZED;
-  }
-}
-
-export function projectMemberRoleToJSON(object: ProjectMemberRole): string {
-  switch (object) {
-    case ProjectMemberRole.PROJECT_MEMBER_ROLE_UNSPECIFIED:
-      return "PROJECT_MEMBER_ROLE_UNSPECIFIED";
-    case ProjectMemberRole.PROJECT_MEMBER_ROLE_ADMIN:
-      return "PROJECT_MEMBER_ROLE_ADMIN";
-    case ProjectMemberRole.PROJECT_MEMBER_ROLE_VIEWER:
-      return "PROJECT_MEMBER_ROLE_VIEWER";
-    case ProjectMemberRole.UNRECOGNIZED:
-    default:
-      return "UNRECOGNIZED";
-  }
-}
 
 /** ProjectServiceListMembersRequest contains the information needed to list members of a project */
 export interface ProjectServiceListMembersRequest {
