+++ app/controlplane/api/gen/jsonschema/controlplane.v1.ProjectInfo.jsonschema.json
@@ -35,8 +35,23 @@
       "type": "string"
     },
     "role": {
-      "description": "Role of the group in the project (admin or viewer)",
-      "type": "string"
+      "anyOf": [
+        {
+          "enum": [
+            "PROJECT_MEMBER_ROLE_UNSPECIFIED",
+            "PROJECT_MEMBER_ROLE_ADMIN",
+            "PROJECT_MEMBER_ROLE_VIEWER"
+          ],
+          "title": "Project Member Role",
+          "type": "string"
+        },
+        {
+          "maximum": 2147483647,
+          "minimum": -2147483648,
+          "type": "integer"
+        }
+      ],
+      "description": "Role of the group in the project (admin or viewer)"
     }
   },
   "title": "Project Info",
